Abstract common logic
